Pool - Tokens with prices are still displayed to enter prices
